requirements.txt
asgiref==3.6.0
bootstrap4==0.1.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
mypy==1.0.1
mypy-extensions==1.0.0
psycopg2==2.9.5
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
